moves and fixes admin auth middleware + changes dev .env path
