+++ src/repositories/AdminRepository.ts
@@ -1,6 +1,5 @@
 import {getCollection} from "./connectToDB";
 import {Admin} from "./schemas";
-import createError from "http-errors";
 
 export default {
     addAdmin(username: string, passwordHash: string) {
@@ -10,7 +9,7 @@ export default {
     async deleteAdmin(username: string) {
         const result = await getCollection("admin").deleteOne({username})
         if (result.deletedCount !== 1) {
-            throw createError(500, "Unable to delete Admin")
+            throw Error("Could not delete Admin")
         }
     },
